chore: switched to binary mode
