+++ crates/cli/Cargo.toml
@@ -26,3 +26,4 @@ anyhow = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing = { workspace = true }
 subprocess = { workspace = true }
+relm4 = { workspace = true }
