+++ crates/daemon/Cargo.toml
@@ -24,3 +24,4 @@ nom = { workspace = true }
 crash-event = { path = "../crash-event" }
 relago-journal-ext ={ path = "../relago-journal-ext"}
 notify ={ path = "../notify"}
+relm4.workspace = true
