+++ crates/daemon/src/journal.rs
@@ -8,12 +8,10 @@ use tracing::error;
 
 use crate::crash::{CoredumpCrash, Crash, OomCrash, ServiceFailureCrash};
 use crate::registry::PluginRegistry;
-use notify::{modal, Modal};
 
 #[derive(Debug)]
 enum ReportRes {
-    Done
-
+    Done,
 }
 
 pub fn run() -> anyhow::Result<()> {
@@ -46,10 +44,10 @@ pub fn run() -> anyhow::Result<()> {
 
     registry.install_filters(&mut journal)?;
 
-    let (tx, rx) = mpsc::channel();
+    // let (tx, rx) = mpsc::channel();
 
     // let mut value = tx.send(ReportRes::Done);
-    let sender = tx.clone();
+    // let sender = tx.clone();
     loop {
         match journal.next() {
             // 0 means "no new entries yet" — block until journald wakes us.
@@ -59,7 +57,6 @@ pub fn run() -> anyhow::Result<()> {
                     .map_err(|e| anyhow!("journal wait failed: {e}"))?;
             }
 
-            
             Ok(_) => match registry.run(&mut journal) {
                 Some(Crash::Coredump(ref r)) => {
                     let unit = "".to_string();
@@ -72,38 +69,21 @@ pub fn run() -> anyhow::Result<()> {
                     };
                     println!("Core dumped: {:?}", r);
                     // handle_crash(cr)?
-                    let value = sender.clone();
-                    thread::spawn(move || {
-                        println!("Handler called inside thread");
+                    // let value = sender.clone();
+                    // relm4::spawn_blocking(move || {
+                    //     println!("Handler called inside thread");
 
-                        match modal(rr){
-                            Some(_) => {
-                                println!("worked");
-                                value.clone().send(ReportRes::Done).unwrap()
-                            }
-                            None => {
-                                panic!("Fuck")
-                            }
-                        }
-                        
-                    });
-
-                    loop {
-                        // thread::sleep(Duration::new(3, 0));
-                        match rx.recv() {
-                            Ok(ReportRes::Done) => {
-                                println!("Exiting worker thread");
-                                break;
-                            }
-                            // Ok(_) => {
-                                // thread::sleep(Duration::new(1, 0));
-                            // }
-                            Err(_) => {
-                                println!("Channel closed");
-                                break;
-                            }
-                        }
-                    }
+                    //     match modal(rr) {
+                    //         Some(_) => {
+                    //             println!("worked");
+                    //             value.clone().send(ReportRes::Done).unwrap()
+                    //         }
+                    //         None => {
+                    //             panic!("Fuck")
+                    //         }
+                    //     }
+                    // });
+                    modal(rr);
                 }
 
                 Some(Crash::ServiceFailure(r)) => {
