+++ crates/notify/Cargo.toml
@@ -7,15 +7,15 @@ license.workspace = true
 homepage.workspace = true
 repository.workspace = true
 
-[lib]
+[[bin]]
 name = "notify"
-path = "src/lib.rs"
+path = "src/main.rs"
 
 [dependencies]
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 notify-rust = "4"
-relm4 = {version = "0.9.1", features=["macros", "libadwaita", "gnome_47"]}
+relm4.workspace = true
 relm4-components = "0.9.1"
 reqwest = { version = "0.13.2", features = [ "multipart", "blocking" ]}
 anyhow.workspace = true
