+++ crates/notify/src/window.rs
@@ -1,19 +1,19 @@
+use std::path::PathBuf;
+
 use relm4::{
     adw::{self, prelude::*},
     component::*,
-    gtk::{self, prelude::*},
-    main_application, *,
+    gtk::{self},
+    RelmApp, Sender,
 };
 use reqwest::blocking::multipart;
 
 use super::Modal;
-use serde_json::{json, Value};
 
 struct AppModel {
     error: Modal,
     title: String,
     state: AppState,
-    url: String,
 }
 
 #[derive(Debug)]
@@ -97,7 +97,7 @@ impl AsyncComponent for AppModel {
                             set_width_request: 180,
                         },
                         gtk::Label  {
-                            set_text: "Xatoliklar bilan bog'liq barcha ma'lumotlar yig'ilmoqda. Iltimos kuting",
+                            set_text: "Ma'lumotlar yig'ilmoqda",
                             set_valign: gtk::Align::Center,
                         },
                     },
@@ -106,7 +106,7 @@ impl AsyncComponent for AppModel {
                         set_hexpand: true,
                         set_vexpand: true,
                         gtk::Label  {
-                            set_text: "Logs sent successfully"
+                            set_text: "xuynya"
                         },
                     },
                 },
@@ -123,7 +123,6 @@ impl AsyncComponent for AppModel {
             error: error.clone(),
             title: format!("Error on {}", error.unit),
             state: AppState::Init,
-            url: "http://localhost:5678".to_string(),
         };
 
         let widgets = view_output!();
@@ -140,20 +139,6 @@ impl AsyncComponent for AppModel {
         match msg {
             AppMsg::Report => {
                 println!("Hello chigga");
-                self.title = "Spinner".to_string();
-                self.state = AppState::Spinning;
-
-                // FIXME: we need to set tmp directory, config path in env
-                match report::create_report("tmp", Some("~/.config/nix"), None) {
-                    Ok(rep_file) => {
-                        // FIXME: we need to set tmp directory, config path in env
-                        match report_srv(rep_file.file.to_str().unwrap()) {
-                            Ok(rep) => self.state = AppState::Send,
-                            Err(_) => self.state = AppState::Init,
-                        }
-                    }
-                    Err(_) => self.state = AppState::Init,
-                }
             }
         }
     }
@@ -164,12 +149,10 @@ pub fn open(error: Modal) {
     app.run_async::<AppModel>(error);
 }
 
-fn report_srv(file_path: &str) -> anyhow::Result<()> {
-    let url = "http://localhost:4242/upload/report";
+fn rep_srv(file: PathBuf) -> anyhow::Result<()> {
+    let url = "http://localhost:5678/upload/report";
 
-    let form = multipart::Form::new()
-        // .text("username", "seanmonstar")
-        .file("report", file_path)?;
+    let form = multipart::Form::new().file("logs", file)?;
 
     let client = reqwest::blocking::Client::new();
     client.post(url).multipart(form).send()?;
